backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.1
cryptography==42.0.0

# Rate Limiting
slowapi==0.1.9

# Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
pdfplumber==0.10.3

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
imbalanced-learn==0.12.0
joblib==1.3.2

# AI / LLM
mistralai>=0.4.0
langchain==0.1.4
langchain-anthropic==0.1.0
langgraph==0.0.19
anthropic==0.18.1

# Redis
redis==5.0.1

# Twilio (Voice + WhatsApp)
twilio>=8.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
ujson==5.9.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Code Quality
black==24.1.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
lxml==5.1.0